Big changes

+++ src/Excel.php
@@ -5,7 +5,8 @@ namespace Chernogolov\Fogcms;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Rap2hpoutre\FastExcel\FastExcel;
-use Chernogolov\Fogcms;\Records;
+use Chernogolov\Fogcms;
+use Chernogolov\Fogcms\Records;
 
 class Excel extends Model
 {
@@ -15,10 +16,21 @@ class Excel extends Model
         $newdata = array();
         $filename = 'file.xlsx';
 
+
         foreach($data as $item)
             $newdata[] = (array) $item;
 
-        unlink($filename);
+        if(file_exists($filename))
+            unlink($filename);
+
+        if(count($newdata) == 0)
+        {
+            $fields = Attr::getFields($node);
+            $newdata[] = array();
+            foreach($fields['default_fields'] as $k => $v)
+                $newdata[0][$k] = "";
+
+        }
         $newdata = Collect($newdata);
         (new FastExcel($newdata))->export($filename);
         
@@ -29,7 +41,7 @@ class Excel extends Model
     {
         !is_array($reg_ids) ? $reg_ids = array($reg_ids) : null;
         $collection = (new FastExcel)->import($file);
-
+        Records::$import = true;
         Records::updateReg($reg_ids, $collection);
     }
 }
